// src/doctor/enums/doctor-specialization.enum.ts
export enum DoctorSpecialization {
    GENERAL = 'General Physician',
    CARDIOLOGIST = 'Cardiologist',
    DERMATOLOGIST = 'Dermatologist',
    NEUROLOGIST = 'Neurologist',
    PEDIATRICIAN = 'Pediatrician',
    ORTHOPEDIC = 'Orthopedic',
    PSYCHIATRIST = 'Psychiatrist',
    ENT = 'ENT Specialist',
    GYNECOLOGIST = 'Gynecologist',
}